Added a unit converter to goat,
Usage:
convert <number>  <multiplier><unit> to <multiplier><unit>
--Convert one unit to another type of unit. The multiplier can be kilo, mega,
pico, etc, and is optional. The number is the amount to be converted.

findunit <unit>
--Searches all the unit names for the supplied string, displaying matches in 
a categorised list

describeunit <unit>
--Displays some information about a unit type.

Also modifed CurrencyConverter so that the command "convert rates" works -
a command I didn't even realise that the module had. Think  stole it from
somewhere.