Confessions 2, initial checkin.
Basic functionality in place, still lots of tidying to do.

Note that you won't have any confessions in your db out of the box; the unit test expects a confessions.gz file to slurp up, after which you're off and running.  Confessions.gz is just bc's confessions.bz2, unbzipped and then gzipped.  If it weren't so hurkin', I'd put it in svn, but that's life.