Relay module for Google's chatbot, Guru.
At present guru doesn't seem to do anything that goat doesn't already do, except for the painless unit conversions.  Which is enough to justify the effort.  There's supposed to be sports scores, which would be nice, but those don't seem to be working atm.  The weather command doesn't give as much info as goat, but it does provide a brief forecast, which is nice.

For more, ask goat:

guru help

The Jabber client library (smack) is a slight pain in that it has a much-used class called Message, so full names are in order.